refactor: :recycle: refactor Members pages and associated components
