+++ my-app/src/Routes.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import "./App.css";
-import Home from "./core/Home";
-// import Members from "./core/Members";
+import Home from "./pages/Home";
+import Members from "./pages/Members";
 // import Event from "./core/Event";
 // import Contact from "./core/Contact";
 // import About from "./core/About";
@@ -15,8 +15,8 @@ const Routes = () => {
       <Router>
         <Switch>
           <Route path="/" exact component={Home} />
-          {/* <Route path='/members' exact component={Members} />
-          <Route path='/alumni' exact component={Alumni} />
+          <Route path='/members' exact component={Members} />
+          {/* <Route path='/alumni' exact component={Alumni} />
           <Route path='/events' exact component={Event} />
           <Route path='/contact' exact component={Contact} />
           <Route path='/about' exact component={About} />
